Create caption_dl_and_clean_wo_set.py
This version should do the same thing as the other BUT now I compare each line only to the last line seen.  This could be helpful when a line of text might be repeated but it's not a .vtt issue but actually important text like repeated lines in a song.